+++ include/reactphysics3d/containers/Deque.h
@@ -55,10 +55,10 @@ class Deque {
         // -------------------- Constants -------------------- //
 
         /// Number of items in a chunk
-        const uint CHUNK_NB_ITEMS = 17;
+        const uint8 CHUNK_NB_ITEMS = 17;
 
         /// First item index in a chunk
-        const uint CHUNK_FIRST_ITEM_INDEX = CHUNK_NB_ITEMS / 2;
+        const uint8 CHUNK_FIRST_ITEM_INDEX = CHUNK_NB_ITEMS / 2;
 
         // -------------------- Attributes -------------------- //
