+++ include/reactphysics3d/containers/Set.h
@@ -348,7 +348,7 @@ class Set {
         // -------------------- Methods -------------------- //
 
         /// Constructor
-        Set(MemoryAllocator& allocator, size_t capacity = 0)
+        Set(MemoryAllocator& allocator, int capacity = 0)
             : mNbUsedEntries(0), mNbFreeEntries(0), mCapacity(0), mBuckets(nullptr),
               mEntries(nullptr), mAllocator(allocator), mFreeIndex(-1) {
 
@@ -446,7 +446,7 @@ class Set {
                 }
             }
 
-            size_t entryIndex;
+            int entryIndex;
 
             // If there are free entries to use
             if (mNbFreeEntries > 0) {
